[#60]rename: expr_util -> util로 밖으로 뺌
