+++ app/visualize/analysis/stmt/expr/parser/list_expr.py
@@ -1,5 +1,5 @@
-from app.visualize.analysis.stmt.expr.expr_util import util
 from app.visualize.analysis.stmt.expr.models.expr_obj import ListObj, ExprObj
+from app.visualize.utils.util import Util
 
 
 class ListExpr:
@@ -18,7 +18,7 @@ class ListExpr:
     @staticmethod
     def _concat_expression(elts: list[ExprObj]):
         elts_expression_lists = [elt.expressions for elt in elts]
-        transposed_expression_lists = util.transpose_with_last_fill(elts_expression_lists)
+        transposed_expression_lists = Util.transpose_with_last_fill(elts_expression_lists)
 
         # [[1, 2, 3], [4, 5, 6]] -> ("[1, 4]", "[2, 5]", "[3, 6]")
         return tuple(f"[{','.join(map(str, t))}]" for t in transposed_expression_lists)
